refactor(library): 拆分观看历史视图职责
